backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
numpy
gradio_client
python-dotenv
opencv-python
google-api-python-client
httpx
